AHCI: Fix read access beyond 4GB for CD/DVD drives
svn:sync-xref-src-repo-rev: r72786

+++ src/VBox/Devices/Storage/DevAHCI.cpp
@@ -1,4 +1,4 @@
-/* $Id: DevAHCI.cpp 37690 2011-06-29 16:05:54Z klaus.espenlaub@oracle.com $ */
+/* $Id: DevAHCI.cpp 37887 2011-07-12 10:50:43Z alexander.eichner@oracle.com $ */
 /** @file
  * VBox storage devices: AHCI controller device (disk and cdrom).
  *                       Implements the AHCI standard 1.1
@@ -4141,13 +4141,13 @@ static int atapiReadSectors(PAHCIPort pAhciPort, PAHCIPORTTASKSTATE pAhciPortTas
     switch (cbSector)
     {
         case 2048:
-            pAhciPortTaskState->uOffset = iATAPILBA * cbSector;
+            pAhciPortTaskState->uOffset = (uint64_t)iATAPILBA * cbSector;
             pAhciPortTaskState->cbTransfer = cSectors * cbSector;
             break;
         case 2352:
         {
             pAhciPortTaskState->pfnPostProcess = atapiReadSectors2352PostProcess;
-            pAhciPortTaskState->uOffset = iATAPILBA * 2048;
+            pAhciPortTaskState->uOffset = (uint64_t)iATAPILBA * 2048;
             pAhciPortTaskState->cbTransfer = cSectors * 2048;
             break;
         }
